fix: Move sources from ui/views/cocoa to ui/views_bridge_mac
https://chromium-review.googlesource.com/1247526
